Generalize commit drag element
