+++ app/src/lib/app-state.ts
@@ -39,7 +39,7 @@ import { IStashEntry } from '../models/stash-entry'
 import { TutorialStep } from '../models/tutorial-step'
 import { UncommittedChangesStrategy } from '../models/uncommitted-changes-strategy'
 import { CherryPickFlowStep } from '../models/cherry-pick'
-import { DragElement } from '../models/drag-element'
+import { DragElement } from '../models/drag-drop'
 import { ILastThankYou } from '../models/last-thank-you'
 
 export enum SelectionType {
